Clean up messy test structure
- Consolidated all database tests into test_database.py
- Renamed test files to follow proper conventions:
  - test_core_simple.py → test_core.py
  - test_cli_simple.py → test_cli.py
  - test_simple_migrations.py → test_migrations.py
- Deleted redundant test files:
  - test_simple.py
  - test_database_extended.py
  - test_more_coverage.py
  - test_integration.py
- Fixed import in conftest.py to use correct module path

Result: 36 passing tests with a clean, organized structure
