+++ tests/test_cli.py
